Refactor: Modify `updateTaskAssignees` to accept `FieldValue`
The `updateTaskAssignees` method has been updated to accept a `FieldValue` parameter instead of a `List<String>`. This change allows for more direct use of Firestore's `arrayUnion` and `arrayRemove` operations.

The `addAssignee` and `removeAssignee` methods have been refactored to utilize this new approach, simplifying the logic for updating task assignees by removing the need to manually manage the list of user IDs on the client side.